requirements.txt
Flask
tensorflow
numpy
pillow
mediapipe
opencv-python
